% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_tracking_polls.R
\name{add_tracking_poll_info}
\alias{add_tracking_poll_info}
\title{Add information on tracking polls}
\usage{
add_tracking_poll_info(x, overlap_days = 1L)
}
\arguments{
\item{x}{a polls \code{data.frame} with \code{collectPeriodTo} and \code{collectPeriodFrom}.}

\item{overlap_days}{the number of days that can overlap between polls without consider it an overlap. 
Defaults to 1L.}
}
\value{
\code{overlap_previous_poll}: the poll overlaps with previous polls
\code{tracking_poll}: the poll part of a tracking poll
\code{tracking_poll_id}: id variable for tracking and ordinary polls
\code{tracking_poll_n}: the number of polls part of the tracking poll
\code{tracking_poll_no}: the tracking poll number
}
\description{
Add information on tracking polls
}
\details{
Identifies tracking polls (polls with overlapping collection periods)
}
