% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_tracking_polls.R
\name{assert_polls}
\alias{assert_polls}
\title{Assert a polls \code{data.frame}}
\usage{
assert_polls(x, columns_to_check = NULL)
}
\arguments{
\item{x}{the object to assess}

\item{columns_to_check}{what columns should be specifically checked. Defaults to none.}
}
\description{
Assert a polls \code{data.frame}
}
