% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_unique_id_by_order.R
\name{create_unique_id_by_order}
\alias{create_unique_id_by_order}
\title{Create a unique id per poll}
\usage{
create_unique_id_by_order(x, use_colnames = c("PublYearMonth", "Company"))
}
\arguments{
\item{x}{a polls data.frame}

\item{use_columns}{which columns should be used to create the unique ID}
}
\description{
The function create a unique ID per poll based on the columns chosen.
If there are multiple polls with the given IDs, the IDs are added with the 
order and total polls in \code{x}.
}
