% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elections.R
\docType{data}
\name{elections}
\alias{elections}
\title{Elections results in Sweden}
\format{
A data frame with one row per election and the same variables as in the Polls dataset.
}
\usage{
elections
}
\description{
A dataset containing the election results during the period of the Swedish polls.
}
\keyword{datasets}
