% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_polls.R
\name{get_polls}
\alias{get_polls}
\alias{get_polls_local}
\alias{read_polls_csv}
\title{Get the Swedish Polls dataset}
\usage{
get_polls(as = "data_frame")

get_polls_local(as)

read_polls_csv(path)
}
\arguments{
\item{as}{How to get polls (raw or as parsed data_frame).}

\item{path}{Local path to Polls.csv}
}
\description{
Download and read in the polls dataset in R as a \code{tibble}.
}
\keyword{Internal}
