% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_tracking_polls.R
\name{handle_tracking_polls}
\alias{handle_tracking_polls}
\title{Handle tracking polls}
\usage{
handle_tracking_polls(
  polls = NULL,
  correction = "use_last_no_overlap",
  verbose = FALSE,
  overlap_days = 1L,
  ...
)
}
\arguments{
\item{polls}{a polls \code{data.frame}}

\item{correction}{what correction to use. 
See \code{supported_tracking_polls_corrections()} for details on what
corrections that can be used.}

\item{verbose}{should information on the handling be printed as messages? Defaults to \code{FALSE}.}

\item{...}{further arguments supplied to the tracking polls correction function(s).}
}
\description{
Handle tracking polls
}
\details{
Some polls overlap in time, especially close to an election. These polls
are usually a way for a house to track parties leading up to an election.
Although, they are commonly using the same observations multiple times
which might introduce problem when using the polls in a model. Hence,
this function helps handling these tracking polls.
}
