% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_polls.R
\name{order_polls}
\alias{order_polls}
\title{Order a polls tbl_df}
\usage{
order_polls(x, decreasing = TRUE)
}
\arguments{
\item{x}{a polls tbl_df extracted by get_polls()}

\item{decreasing}{logical. Should the sort order be increasing or decreasing?}
}
\description{
Order a polls data table by year and month
}
