% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_tracking_polls.R
\name{supported_tracking_polls_corrections}
\alias{supported_tracking_polls_corrections}
\title{Implemented corrections for tracking polls}
\usage{
supported_tracking_polls_corrections()
}
\description{
Implemented corrections for tracking polls
}
