% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_tracking_polls.R
\name{use_last_no_overlap}
\alias{use_last_no_overlap}
\title{Use the last poll without overlap with previous polls}
\usage{
use_last_no_overlap(x, overlap_days)
}
\arguments{
\item{x}{a polls \code{data.frame}}

\item{overlap_days}{the number of days that can overlap between polls without consider it an overlap.}
}
\description{
Use the last poll without overlap with previous polls
}
