% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_tracking_polls.R
\name{use_last_no_overlap_indicator}
\alias{use_last_no_overlap_indicator}
\title{Return an indicator variable for last polls without overlap}
\usage{
use_last_no_overlap_indicator(polls, overlap_days = 1L)
}
\arguments{
\item{polls}{a polls object with variables \code{house}, \code{collectPeriodFrom}, and \code{collectPeriodTo}.}

\item{overlap_days}{the number of days that can overlap between polls without consider it an overlap.}
}
\description{
Return an indicator variable for last polls without overlap
}
